UPSTREAM: drm/i915/bxt: add support for setting backlight freq from vbt
The only missing piece is the function to convert frequency to PWM
register value. The PWM is based on 19.2 MHz clock, except for BXT A
step, which is based on CDCLK, and which we ignore.

Reviewed-by: Imre Deak <imre.deak@intel.com>
Tested-by: Imre Deak <imre.deak@intel.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1448958232-26520-1-git-send-email-jani.nikula@intel.com
(cherry picked from commit 2dd6982e471f68137afd864fb4de64c59c7f5388
from v4.5 )
Signed-off-by: Tarun Vyas <tarun.vyas@intel.com>

Change-Id: Ia16ff5a15d1d9f3e3bac3175e17089aa283774d1