BACKPORT: ASoC: Intel: Skylake: Add support for active suspend
Some of the usecases can be marked as 'ignore_suspend' by
machine. For these on suspend we should keep audio controller
ON by saving the state and not suspending the device

For this we need to maintain a counter for these streams and be
active on suspend when such a stream is opened.

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 4557c305d4fc9356563a1d41fa6fe29e494f0460)
Signed-off-by: Ilkka Koskinen <ilkka.koskinen@intel.com>

Conflicts:
	sound/soc/intel/skylake/skl.h

Change-Id: Id9dc9714714220eb419aa3851ed7b3c6e2398300