UPSTREAM: drm/i915: Sanity check DP AUX message buffer and size
While we are calling intel_dp_aux_transfer() with msg->size=0 whenever
msg->buffer is NULL, passing NULL to memcpy() is undefined according to
the ISO C standard. I haven't found any notes about this in the GNU C's
or the kernel's documentation of the function and can't imagine what it
would do with the NULL ptr. To better document this use of the
parameters it still make sense to add an explicit check for this to the
code.

Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: David Weinehall <david.weinehall@intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1454071949-24677-1-git-send-email-imre.deak@intel.com
(cherry picked from commit d81a67cc1bfdc3efed8590ee4e07a40b198908ba
from v4.6-rc2 )
Signed-off-by: Tarun Vyas <tarun.vyas@intel.com>

Change-Id: I7fdcaa3324e8c2c63d53acb2b842b22a58936e18