UPSTREAM: drm/exynos: dp: Fix panel and bridge lookup logic
Commit a9fa852886fd ("drm/exynos: dp: add of_graph dt binding support
for panel") made the Exynos DP DT binding more consistent since the OF
graph could be used to lookup either a panel or a bridge device node.

Before that commit, a panel would be looked up using a phandle and a
bridge using the OF graph which made the DT binding not consistent.

But the patch broke the later case since not finding a panel dev node
would cause the driver's to do a probe deferral instead of attempting
to lookup a bridge device node associated with the remote endpoint.

So instead of returning a -EPROBE_DEFER if a panel is not found, check
if there's a bridge and only do a probe deferral if both aren't found.

BUG=chrome-os-partner:50756
TEST=Build and boot on rk3288

Change-Id: I8bebfab4a17099ccae5a54249dd80d76f1b46242
Signed-off-by: Javier Martinez Canillas <javier@osg.samsung.com>
Tested-by: Michal Suchanek <hramrach@gmail.com>
Signed-off-by: Inki Dae <inki.dae@samsung.com>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 37e110625eeeaba83e8cb763ab7645f0678c6f8e)