UPSTREAM: ALSA: hda - build chmap kctl based on pcm in hdmi audio
Build chmap kctl based on pcm. The first chmap kctl will be
mapped to the first pcm, and so on.

When a monitor is connected to a pin, the chmap kctl can
find the pin and the monitor through the pcm index.

Signed-off-by: Libin Yang <libin.yang@linux.intel.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 022f344b41a54a995d90450726842518e91aec29)
Signed-off-by: Ilkka Koskinen <ilkka.koskinen@intel.com>