UPSTREAM: mei: fix double freeing of a cb during link reset
Fix double freeing of the cb that can happen if link reset kicks  in the
middle of blocked write from a device on the cl bus.

Free cb inside mei_cl_write function on failure and drop cb free
operation from callers, during a link reset the mei_cl_write function
returns with an error,  but the caller doesn't know if the cb was
already queued or not so it doesn't know if the cb will be freed upon
queue reclaim or it has to free it itself.

Signed-off-by: Alexander Usyskin <alexander.usyskin@intel.com>
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Signed-off-by: Ilkka Koskinen <ilkka.koskinen@intel.com>
(cherry picked from commit 6cbb097fd3903fd6b419303ee8dc7f72b47d06f0)
Change-Id: Ie03a5e2cf6945bb733de26e3bafc4f82832851be