UPSTREAM: ASoC: Intel: Skylake: Fix for unloading module only when it is loaded
Module needs to be unloaded only when it is loaded successfully.
To fix this, first correct the module state sequence and set module
state to LOADED if module is loaded successfully.
When unloading the module check if module state is not in UNINIT,
then unload it.

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit d643678b9a4e432a574b2ec601216afd720c69b6)
Signed-off-by: Ilkka Koskinen <ilkka.koskinen@intel.com>