UPSTREAM: mmc: block: Pause re-tuning while switched to the RPMB partition
Re-tuning is not possible when switched to the RPMB
partition.  However re-tuning should not be needed
if re-tuning is done immediately before switching,
a small set of operations is done, and then we
immediately switch back to the main partition.

A previous patch ensured that we immediately switch
back to the main partition.

This patch uses the new facility to "pause" re-tuning
before switching to the RPMB partition, and to "unpause"
it after switching from the RPMB partition.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
(cherry picked from commit 57da0c042f4af52614f4bd1a148155a299ae5cd8)
Signed-off-by: Ilkka Koskinen <ilkka.koskinen@intel.com>

Change-Id: I9e87e01f6283c7794f40a7d6e5709b80ec7131f8