UPSTREAM: usb: dwc3: core: re-factor init and exit paths
The idea of this patch is for dwc3_core_init() to
abstract all the details about how to initialize
dwc3 and dwc3_core_exit() to do the same for
teardown.

With this, we can simplify suspend/resume operations
by a large margin and always know that we're going
to start dwc3 from a known starting point.

Signed-off-by: Felipe Balbi <felipe.balbi@linux.intel.com>
(cherry picked from commit c499ff71ff2a281366c6ec7a904c547d806cbcd1)
Signed-off-by: Ilkka Koskinen <ilkka.koskinen@intel.com>

Change-Id: I64607cb459e98357c1e7aeae275bb961f0444223