UPSTREAM: ASoC: Intel: Skylake: Add strip extended manifest utility
Some upcoming platforms like broxton etc have extended manifest
in firmware binary. This is not required to be downloaded to DSP.
So driver needs to strip this before downloading.

Add a utility function to check if a header exists, and remove it
in that case

Signed-off-by: Ramesh Babu <ramesh.babu@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 6eee87261f69e366bfe9b908ae3b441efb8e28ea)
Signed-off-by: Ilkka Koskinen <ilkka.koskinen@intel.com>

Change-Id: I5940164a9016e7323a758e8908bff3851326c41c