WA: Fix bluetooth service restart.
If the BT hardware enables a window size greater than 1, the
Bluetooth HAL can send more than one command at a time. The Android HAL
expects a command complete event for each command it sends to the driver.

For btusb driver. the firmware will sometimes send one event complete for the
HCI_OP_WRITE_SCAN_ENABLE command even if two commands are sent at the same time,
this causes the HAL to timeout and to complain about the missing event.
The timeout error is fatal and will cause the BT subsystem to loop/restart.

This patch will restrict the Joule HAL to send one command at a time by
having the kernel force the returned window size to 1.

BUG: 34691331
Change-Id: I48d085a5bb14d6d25bef85bffe93081b96879b49
Signed-off-by: mabbas <mohamed.abbas@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>